chore: moved game commands to the textcommand folder
