+++ src/main/java/org/virep/jdabot/textcommands/InventoryCommand.java
@@ -1,10 +1,12 @@
-package org.virep.jdabot.commands.game;
+package org.virep.jdabot.textcommands;
 
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.Guild;
+import net.dv8tion.jda.api.entities.Message;
 import net.dv8tion.jda.api.entities.MessageEmbed;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
+import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import net.dv8tion.jda.api.interactions.commands.build.Commands;
 import net.dv8tion.jda.api.interactions.commands.build.SlashCommandData;
 import net.dv8tion.jda.api.interactions.components.selections.StringSelectMenu;
@@ -12,8 +14,8 @@ import org.json.JSONObject;
 import org.json.JSONTokener;
 import org.virep.jdabot.commands.games.ActivityCommand;
 import org.virep.jdabot.database.Database;
+import org.virep.jdabot.handlers.TextCommand;
 import org.virep.jdabot.language.Language;
-import org.virep.jdabot.handlers.SlashCommand;
 import org.virep.jdabot.utils.ErrorManager;
 
 import java.awt.*;
@@ -27,23 +29,10 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-public class InventoryCommand implements SlashCommand {
+public class InventoryCommand implements TextCommand {
     @Override
     public String getName() {
         return "inventory";
-    }
-
-    @Override
-    public SlashCommandData getCommandData() {
-        return Commands.slash(getName(), "Displays your Game Inventory");
-    }
-
-    @Override
-    public List<Permission> getBotPermissions() {
-        List<Permission> perms = new ArrayList<>();
-        Collections.addAll(perms, Permission.MESSAGE_EMBED_LINKS);
-
-        return perms;
     }
 
     @Override
@@ -51,18 +40,20 @@ public class InventoryCommand implements SlashCommand {
         return true;
     }
 
+
     //TODO: Don't display resources the player doesn't have
     @Override
-    public void execute(SlashCommandInteractionEvent event) {
+    public void execute(MessageReceivedEvent event, String[] args) throws SQLException {
         Guild guild = event.getGuild();
+        Message message = event.getMessage();
 
         try (Connection connection = Database.getConnection(); PreparedStatement statement = connection.prepareStatement("SELECT * FROM inventory WHERE userID = ?")) {
-            statement.setString(1, event.getUser().getId());
+            statement.setString(1, event.getAuthor().getId());
 
             ResultSet result = statement.executeQuery();
 
             if (!result.first()) {
-                event.reply(Language.getString("INVENTORY_NOT_REGISTERED", guild)).queue();
+                message.reply(Language.getString("INVENTORY_NOT_REGISTERED", guild)).queue();
                 return;
             }
 
@@ -72,7 +63,7 @@ public class InventoryCommand implements SlashCommand {
             JSONObject itemObject = new JSONObject(tokener);
 
             MessageEmbed embed = new EmbedBuilder()
-                    .setTitle(String.format(Language.getString("INVENTORY_EMBED_GENERAL_TITLE", guild), event.getUser().getName()))
+                    .setTitle(String.format(Language.getString("INVENTORY_EMBED_GENERAL_TITLE", guild), event.getAuthor().getName()))
                     .setColor(Color.RED)
                     .addField("General", Language.getString("INVENTORY_EMBED_GENERAL_FIELD_ONE", guild).formatted(result.getInt("level"), result.getInt("xp"), 0, result.getInt("mana"), 100, result.getInt("hp"), 100), true)
                     .addField("Items", Language.getString("INVENTORY_EMBED_GENERAL_FIELD_ITEMS", guild).formatted(
@@ -94,9 +85,9 @@ public class InventoryCommand implements SlashCommand {
                     .setMinValues(1)
                     .build();
 
-            event.replyEmbeds(embed).addActionRow(selectMenu).queue();
+            message.replyEmbeds(embed).addActionRow(selectMenu).queue();
         } catch (SQLException e) {
-            ErrorManager.handle(e, event);
+            throw new SQLException(e);
         }
     }
 }
